LayoutTests/streams/readablestream-cancel-after-waiting-for-chunks-expected.txt

ReadableStream cancellation puts the stream in a closed state (after waiting for chunks).

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS ready promise vended before the cancellation should fulfill
PASS rs.state is "closed"
PASS ready promise vended before the cancellation should fulfill
PASS closed promise vended before the cancellation should fulfill
PASS closed promise vended after the cancellation should fulfill
PASS ready promise vended after the cancellation should fulfill
PASS successfullyParsed is true

TEST COMPLETE
